restructured poll queries mutaion
