+++ src/providers/queries/postQueries/pollQueries.ts
@@ -1,36 +1,10 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import QUERIES from "../queryKeys";
-import { getPollData } from "../../polls/polls";
+import { castPollVote, getPollData } from "../../polls/polls";
 import { PostMetadata } from "../../hive/hive.types";
 import { useMemo, useState } from "react";
-import { getActiveKey, getDigitPinCode, sendHiveOperations } from "../../hive/dhive";
-import { Operation, PrivateKey } from "@esteemapp/dhive";
 import { Poll } from "../../polls/polls.types";
 import { useAppSelector } from "../../../hooks";
-
-
-const executePollAction = (id: string, json: any, currentAccount: any, pinHash: string) => {
-    const pin = getDigitPinCode(pinHash);
-    const key = getActiveKey(currentAccount.local, pin);
-    const username = currentAccount.name;
-
-    if (key) {
-        const privateKey = PrivateKey.fromString(key);
-
-        const op = {
-            id,
-            json: JSON.stringify(json),
-            required_auths: [username],
-            required_posting_auths: [],
-        };
-        const opArray: Operation[] = [['custom_json', op]];
-        return sendHiveOperations(opArray, privateKey);
-    }
-
-    return Promise.reject(
-        new Error('Check private key permission! Required private active key or above.'),
-    );
-};
 
 
 
@@ -73,7 +47,7 @@ export const useGetPollQuery = (_author?: string, _permlink?: string, metadata?:
         },
     );
 
-    //TODO: inject poll cache here.
+    //TODO: use injectPollVoteCache here for simplifity and code reuseability
     // const data = useInjectVotesCache(query.data);
 
     return {
@@ -107,17 +81,11 @@ export function useVotePollMutation(poll: Poll | null) {
                 throw new Error("Invalid vote")
             }
 
-            await executePollAction("polls", {
-                poll: poll.poll_trx_id,
-                action: "vote",
-                choice: choiceNum
-            }, currentAccount, pinHash);
+            await castPollVote(poll.poll_trx_id, choiceNum, currentAccount, pinHash)
 
             return { choiceNum };
         },
         onMutate: ({ choiceNum }) => {
-
-
             queryClient.setQueryData<ReturnType<typeof useGetPollQuery>["data"]>(
                 [QUERIES.POST.GET_POLL, poll?.author, poll?.permlink],
                 (data) => {
@@ -125,6 +93,7 @@ export function useVotePollMutation(poll: Poll | null) {
                         return data;
                     }
 
+                    //TOOD: use injectPollVoteCache here for simplifity and code reuseability
                     const existingVote = data.poll_voters?.find((pv) => pv.name === currentAccount!!.username);
                     const previousUserChoice = data.poll_choices?.find(
                         (pc) => existingVote?.choice_num === pc.choice_num
@@ -138,9 +107,9 @@ export function useVotePollMutation(poll: Poll | null) {
                     const otherVoters =
                         data.poll_voters?.filter((pv) => pv.name !== currentAccount!!.username) ?? [];
 
-                   
+
                     let poll_choices = data.poll_choices;
-                    if(previousUserChoice?.choice_num !== choice.choice_num){
+                    if (previousUserChoice?.choice_num !== choice.choice_num) {
                         poll_choices = [
                             ...notTouchedChoices,
                             previousUserChoice
@@ -161,7 +130,6 @@ export function useVotePollMutation(poll: Poll | null) {
                     }
 
 
-
                     return {
                         ...data,
                         poll_choices,
@@ -172,53 +140,64 @@ export function useVotePollMutation(poll: Poll | null) {
                     } as ReturnType<typeof useGetPollQuery>["data"];
                 }
             )
+        },
+
+        onSuccess: (resp) => {
+            //TODO: update cache here
+        },
+        onError: (err) => {
+            //TOOD: reverse mutation here
         }
 
-        //   onSuccess: (resp) =>
-        //     queryClient.setQueryData<ReturnType<typeof useGetPollDetailsQuery>["data"]>(
-        //       [QueryIdentifiers.POLL_DETAILS, poll?.author, poll?.permlink],
-        //       (data) => {
-        //         if (!data || !resp) {
-        //           return data;
-        //         }
 
-        //         const existingVote = data.poll_voters?.find((pv) => pv.name === activeUser!!.username);
-        //         const previousUserChoice = data.poll_choices?.find(
-        //           (pc) => existingVote?.choice_num === pc.choice_num
-        //         );
-        //         const choice = data.poll_choices?.find((pc) => pc.choice_num === resp.choiceNum)!!;
+    });
+}
 
-        //         const notTouchedChoices = data.poll_choices?.filter(
-        //           (pc) => ![previousUserChoice?.choice_num, choice?.choice_num].includes(pc.choice_num)
-        //         );
-        //         const otherVoters =
-        //           data.poll_voters?.filter((pv) => pv.name !== activeUser!!.username) ?? [];
+//TODO: use to create, update and remove poll vote entry from votes data
+const injectPollVoteCache = () => {
+        // queryClient.setQueryData<ReturnType<typeof useGetPollDetailsQuery>["data"]>(
+        //   [QueryIdentifiers.POLL_DETAILS, poll?.author, poll?.permlink],
+        //   (data) => {
+        //     if (!data || !resp) {
+        //       return data;
+        //     }
 
-        //         return {
-        //           ...data,
-        //           poll_choices: [
-        //             ...notTouchedChoices,
-        //             previousUserChoice
-        //               ? {
-        //                   ...previousUserChoice,
-        //                   votes: {
-        //                     total_votes: (previousUserChoice?.votes?.total_votes ?? 0) - 1
-        //                   }
-        //                 }
-        //               : undefined,
-        //             {
-        //               ...choice,
+        //     const existingVote = data.poll_voters?.find((pv) => pv.name === activeUser!!.username);
+        //     const previousUserChoice = data.poll_choices?.find(
+        //       (pc) => existingVote?.choice_num === pc.choice_num
+        //     );
+        //     const choice = data.poll_choices?.find((pc) => pc.choice_num === resp.choiceNum)!!;
+
+        //     const notTouchedChoices = data.poll_choices?.filter(
+        //       (pc) => ![previousUserChoice?.choice_num, choice?.choice_num].includes(pc.choice_num)
+        //     );
+        //     const otherVoters =
+        //       data.poll_voters?.filter((pv) => pv.name !== activeUser!!.username) ?? [];
+
+        //     return {
+        //       ...data,
+        //       poll_choices: [
+        //         ...notTouchedChoices,
+        //         previousUserChoice
+        //           ? {
+        //               ...previousUserChoice,
         //               votes: {
-        //                 total_votes: (choice?.votes?.total_votes ?? 0) + 1
+        //                 total_votes: (previousUserChoice?.votes?.total_votes ?? 0) - 1
         //               }
         //             }
-        //           ].filter((el) => !!el),
-        //           poll_voters: [
-        //             ...otherVoters,
-        //             { name: activeUser?.username, choice_num: resp.choiceNum }
-        //           ]
-        //         } as ReturnType<typeof useGetPollDetailsQuery>["data"];
-        //       }
-        //     )
-    });
+        //           : undefined,
+        //         {
+        //           ...choice,
+        //           votes: {
+        //             total_votes: (choice?.votes?.total_votes ?? 0) + 1
+        //           }
+        //         }
+        //       ].filter((el) => !!el),
+        //       poll_voters: [
+        //         ...otherVoters,
+        //         { name: activeUser?.username, choice_num: resp.choiceNum }
+        //       ]
+        //     } as ReturnType<typeof useGetPollDetailsQuery>["data"];
+        //   }
+        // )
 }
